+++ gwcelery/tasks/orchestrator.py
@@ -271,41 +271,6 @@ def _create_voevent(classification, *args, **kwargs):
     return gracedb.create_voevent(*args, **kwargs)
 
 
-def _get_number_of_instruments(gracedb_id):
-    """Get the number of gravitational-wave instruments that contributed to the
-    ranking statistic of a coincident event.
-
-    Parameters
-    ----------
-    gracedb_id : str
-        The GraceDB ID.
-
-    Returns
-    -------
-    int
-        The number of instruments that contributed to the ranking statistic for
-        the event.
-
-    Notes
-    -----
-    The number of instruments that contributed *data* to an event is given by
-    the ``instruments`` key of the GraceDB event JSON structure. However, some
-    pipelines (e.g. gstlal) have a distinction between which instruments
-    contributed *data* and which were considered in the *ranking* of the
-    candidate. For such pipelines, we infer which pipelines contributed to the
-    ranking by counting only the SingleInspiral records for which the chi
-    squared field is non-empty.
-    """
-    event = gracedb.get_event(gracedb_id)
-    attrib = event['extra_attributes']
-    try:
-        singles = attrib['SingleInspiral']
-    except KeyError:
-        return len(event.get('instruments', '').split(','))
-    else:
-        return sum(single.get('chisq') is not None for single in singles)
-
-
 @app.task(ignore_result=True, shared=False)
 def preliminary_alert(event, superevent_id):
     """Produce a preliminary alert by copying any sky maps.
@@ -341,6 +306,7 @@ def preliminary_alert(event, superevent_id):
         app.conf['preliminary_alert_trials_factor'][event['group'].lower()]
     far_threshold = \
         app.conf['preliminary_alert_far_threshold'][event['group'].lower()]
+    num_ifos = gracedb.get_number_of_instruments(preferred_event_id)
     should_publish = (
         not event['offline']
         and trials_factor * event['far'] <= far_threshold
@@ -348,7 +314,7 @@ def preliminary_alert(event, superevent_id):
         # FIXME: For now, we do not issue public alerts for single-detecetor
         # events. Remove this after we have gotten some more experience with
         # single-instrument events.
-        and _get_number_of_instruments(preferred_event_id) != 1
+        and num_ifos != 1
         # FIXME: For now, disable automated alerts except for cWB and gstlal.
         # Remove this after oLIB, MBTA, PyCBC, and SPIIR have completed review.
         and event['pipeline'].lower() in {'cwb', 'gstlal'})