+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
     to confirm receipt of our own GCN Notices from a managed, private
     connection to an anonymous, public connection.
 
+-   Calculation of number of instruments is now unified across superevent
+    manager and orchestrator using gracedb method ``get_number_of_instruments``.
+
 0.4.1 (2019-04-02)
 ------------------